US #905264 Vulnerabilities Injection via Pipeline as code (Fortify SSC) fixes
